Actualized big chunk of project

+++ app/src/main/java/ru/pnzgu/stringapp/ui/hash/HashViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
 class HashViewModel : ViewModel() {
-    val time = MutableLiveData<Long>()
+    val time = MutableLiveData<Double>()
     val timeAverage = MutableLiveData<Double>()
     val comparisonCount = MutableLiveData<Long>()
     val comparisonAverage = MutableLiveData<Double>()
-    val foundCount = MutableLiveData<Long>()
+    val foundCount = MutableLiveData<Double>()
 }
